+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 * Added percentage type for thresholds
 * Added ``datasource`` parameter to CloudWatch targets
 * Added support for auto panels ids to AlertList panel
+* Added support for fields value in Stat panel
 
 0.6.1 (2021-11-23)
 ==================